+++ components/blocks/LeaderboardInfo/Desktop/Desktop.vue
