+++ components/blocks/LeaderboardInfo/Mobile/Mobile.vue
